backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Pillow==10.3.0
django-filter==24.3
drf-yasg==1.21.7
